requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
mysqlclient==1.4.6
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pydantic==1.5.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.2
urllib3==1.25.9
uvicorn==0.11.3
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
wrapt==1.11.2
